refactor: resolve audit issues #669–#672 — shared utilities, component splits, and dead code removal (#725)
* feat: extract truncateAddress to shared lib/format utility

* refactor: use shared truncateAddress in InvoiceCard

* refactor: use shared truncateAddress in InvoiceTable

* refactor: use shared truncateAddress in dashboard page

* refactor: use shared truncateAddress in ConfirmationDialog

* refactor: use shared truncateAddress in WalletConnect

* refactor: use shared truncateAddress in profile page

* refactor: use shared truncateAddress in InvoiceDetailClient

* refactor: use shared getErrorMessage from lib/errors in InvoiceForm

* refactor: use shared getErrorMessage from lib/errors in InvoiceDetailClient

* refactor: use shared getErrorMessage from lib/errors in WalletConnect

* chore: remove empty, unreferenced StatusTimeline component and its test file

* refactor: split DiscountCalculator into SmeCalculator and LpCalculator subcomponents

* feat: add validateDiscountBps utility for unit testing

* refactor: extract inline BPS validation to shared validateDiscountBps in InvoiceCard

* style: fix Prettier formatting in LpCalculator
