+++ apps/web/components/invoice/InvoiceCard.tsx
@@ -21,6 +21,8 @@ import {
   Clock,
 } from "lucide-react";
 import { formatAmount } from "@/lib/assets";
+import { truncateAddress } from "@/lib/format";
+import { validateDiscountBps } from "@/lib/validation";
 import { useConfirmDialogStore } from "@/store/confirmDialog";
 
 interface InvoiceCardProps {
@@ -55,11 +57,6 @@ export const InvoiceCard = React.memo(function InvoiceCard({
   const [showListForm, setShowListForm] = useState(false);
   const discountBpsId = useId();
 
-  const truncateAddr = (addr: string) => {
-    if (!addr) return "";
-    return `${addr.slice(0, 6)}...${addr.slice(-4)}`;
-  };
-
   const copyToClipboard = async (text: string, type: "id" | "buyer") => {
     await navigator.clipboard.writeText(text);
     if (type === "id") {
@@ -76,7 +73,7 @@ export const InvoiceCard = React.memo(function InvoiceCard({
     setError(null);
     try {
       const bps = parseInt(discountBpsInput, 10);
-      if (isNaN(bps) || bps <= 0 || bps > 10000) {
+      if (!validateDiscountBps(bps)) {
         throw new Error(
           "Discount basis points must be between 1 and 10,000 (100%)",
         );
@@ -133,7 +130,7 @@ export const InvoiceCard = React.memo(function InvoiceCard({
         <div className="space-y-1">
           <div className="flex items-center gap-1.5">
             <span className="text-[10px] font-mono text-slate-500 font-bold">
-              INV#{truncateAddr(invoice.id)}
+              INV#{truncateAddress(invoice.id)}
             </span>
             <button
               onClick={(e) => {
@@ -189,7 +186,7 @@ export const InvoiceCard = React.memo(function InvoiceCard({
           <span className="text-slate-500">Buyer Address:</span>
           <div className="flex items-center gap-1">
             <span className="text-slate-300" title={invoice.buyer}>
-              {truncateAddr(invoice.buyer)}
+              {truncateAddress(invoice.buyer)}
             </span>
             <button
               onClick={(e) => {
